requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.6
django-allauth==0.52.0
django-cors-headers==3.13.0
django-countries==7.5.1
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
greenlet==2.0.2
httpie==3.2.1
idna==3.4
markdown-it-py==2.1.0
mdurl==0.1.2
msgpack==1.0.4
multidict==6.0.4
mypy-extensions==1.0.0
neovim==0.3.1
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
phonenumbers==8.13.6
Pillow==9.4.0
platformdirs==3.0.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pynvim==0.4.3
PySocks==1.7.1
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.1
sqlparse==0.4.3
twilio==7.16.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
